Move Policy to def; add field to Formula.Action.
