+++ def/action.go
@@ -8,6 +8,7 @@ type Action struct {
 	Entrypoint []string `json:"command,omitempty"` // executable to invoke as the task.  included in the conjecture.
 	Cwd        string   `json:"cwd,omitempty"`     // working directory to set when invoking the executable.  if not set, will be defaulted to "/".
 	Env        Env      `json:"env,omitempty"`     // environment variables.  included in the conjecture.
+	Policy     Policy   `json:"policy,omitempty"`
 	Escapes    Escapes  `json:"escapes,omitempty"`
 }
